chore: fixes to tests
