+++ packages/cmdl-polymers/src/graph.ts
@@ -88,11 +88,6 @@ export class PolymerGraph {
    * @param property any
    */
   public setNodeProperty(nodePath: string, property: any): void {
-    logger.debug(
-      `Node Property: ${property.name}\nPath: ${[...this.nodes.keys()].join(
-        ", "
-      )}`
-    );
     const node = this.nodes.get(nodePath);
 
     if (!node) {
